+++ packages/colony-js-client/src/ColonyClient/index.js
@@ -22,7 +22,7 @@ import {
 } from '../constants';
 
 type Address = string;
-type PayableAddress = string;
+type TokenAddress = string;
 type Role = $Keys<typeof ROLES>;
 type IPFSHash = string;
 
@@ -136,7 +136,7 @@ export default class ColonyClient extends ContractClient {
     {
       taskId: number, // Integer taskId.
       role: Role, // Role the payout is specified for: MANAGER, EVALUATOR, or WORKER.
-      source: PayableAddress, // Address of the token's contract `0x0` value indicates Ether.
+      token: TokenAddress, // Address of the token's contract. `0x0` value indicates Ether.
     },
     {
       amount: BigNumber, // Amount of specified tokens to payout for that task and a role.
@@ -185,12 +185,12 @@ export default class ColonyClient extends ContractClient {
     ColonyClient,
   >;
   /*
-    Gets a balance for a certain source in a specific pot.
+    Gets a balance for a certain token in a specific pot.
   */
   getPotBalance: ColonyClient.Caller<
     {
       potId: number, // Integer potId.
-      source: PayableAddress, // Address to get funds from, such as the token contract address, or empty address (`0x0` for Ether)
+      token: TokenAddress, // Address to get funds from, such as the token contract address, or empty address (`0x0` for Ether)
     },
     {
       balance: BigNumber, // Balance for token `token` in pot `potId`.
@@ -202,7 +202,7 @@ export default class ColonyClient extends ContractClient {
   */
   getNonRewardPotsTotal: ColonyClient.Caller<
     {
-      source: PayableAddress, // Address of the token's contract `0x0` value indicates Ether.
+      token: TokenAddress, // Address of the token's contract. `0x0` value indicates Ether.
     },
     {
       total: BigNumber, // All tokens that are not within the colony's `rewards` pot.
@@ -220,7 +220,7 @@ export default class ColonyClient extends ContractClient {
       blockNumber: number, // Block number at the time of creation.
       remainingTokenAmount: BigNumber, // Remaining (unclaimed) amount of tokens.
       reputationRootHash: string, // Reputation root hash at the time of creation.
-      source: PayableAddress, // Token address (`0x0` value indicates Ether).
+      token: TokenAddress, // Token address (`0x0` value indicates Ether).
       totalTokenAmountForRewardPayout: BigNumber, // Total amount of tokens taken aside for reward payout.
       totalTokens: BigNumber, // Total colony tokens at the time of creation.
     },
@@ -321,7 +321,7 @@ export default class ColonyClient extends ContractClient {
   setTaskEvaluatorPayout: ColonyClient.MultisigSender<
     {
       taskId: number, // Integer taskId.
-      source: PayableAddress, // Address to send funds from, e.g. the token's contract address, or empty address (`0x0` for Ether)
+      token: TokenAddress, // Address to send funds from, e.g. the token's contract address, or empty address (`0x0` for Ether)
       amount: BigNumber, // Amount to be paid.
     },
     {},
@@ -333,7 +333,7 @@ export default class ColonyClient extends ContractClient {
   setTaskManagerPayout: ColonyClient.Sender<
     {
       taskId: number, // Integer taskId.
-      source: PayableAddress, // Address to send funds from, e.g. the token's contract address, or empty address (`0x0` for Ether)
+      token: TokenAddress, // Address to send funds from, e.g. the token's contract address, or empty address (`0x0` for Ether)
       amount: BigNumber, // Amount to be paid.
     },
     {},
@@ -345,7 +345,7 @@ export default class ColonyClient extends ContractClient {
   setTaskWorkerPayout: ColonyClient.MultisigSender<
     {
       taskId: number, // Integer taskId.
-      source: PayableAddress, // Address to send funds from, e.g. the token's contract address, or empty address (`0x0` for Ether)
+      token: TokenAddress, // Address to send funds from, e.g. the token's contract address, or empty address (`0x0` for Ether)
       amount: BigNumber, // Amount to be paid.
     },
     {},
@@ -418,13 +418,13 @@ export default class ColonyClient extends ContractClient {
     ColonyClient,
   >;
   /*
-    Claims the payout in `source` denomination for work completed in task `taskId` by contributor with role `role`. Allowed only by the contributors themselves after task is finalized. Here the network receives its fee from each payout. Ether fees go straight to the Meta Colony whereas Token fees go to the Network to be auctioned off.
+    Claims the payout for `token` denomination for work completed in task `taskId` by contributor with role `role`. Allowed only by the contributors themselves after task is finalized. Here the network receives its fee from each payout. Ether fees go straight to the Meta Colony whereas Token fees go to the Network to be auctioned off.
   */
   claimPayout: ColonyClient.Sender<
     {
       taskId: number, // Integer taskId.
       role: Role, // Role of the contributor claiming the payout: MANAGER, EVALUATOR, or WORKER
-      source: PayableAddress, // Address to claim funds from, e.g. the token's contract address, or empty address (`0x0` for Ether)
+      token: TokenAddress, // Address to claim funds from, e.g. the token's contract address, or empty address (`0x0` for Ether)
     },
     {},
     ColonyClient,
@@ -456,11 +456,11 @@ export default class ColonyClient extends ContractClient {
     ColonyClient,
   >;
   /*
-    Move any funds received by the colony in `source` denomination to the top-levl domain pot, siphoning off a small amount to the rewards pot. No fee is taken if called against a colony's own token.
+    Move any funds received by the colony for `token` denomination to the top-levl domain pot, siphoning off a small amount to the rewards pot. No fee is taken if called against a colony's own token.
   */
   claimColonyFunds: ColonyClient.Sender<
     {
-      source: PayableAddress, // Address to claim funds from; empty address (`0x0` for Ether)
+      token: TokenAddress, // Address to claim funds from; empty address (`0x0` for Ether)
     },
     {},
     ColonyClient,
@@ -483,7 +483,7 @@ export default class ColonyClient extends ContractClient {
       fromPot: number, // Origin pot Id.
       toPot: number, // Destination pot Id.
       amount: BigNumber, // Amount of funds to move.
-      address: PayableAddress, // Address of the token contract (`0x0` value indicates Ether).
+      token: TokenAddress, // Address of the token contract (`0x0` value indicates Ether).
     },
     {},
     ColonyClient,
@@ -513,7 +513,7 @@ export default class ColonyClient extends ContractClient {
   */
   startNextRewardPayout: ColonyClient.Sender<
     {
-      token: Address, // Address of token used for reward payout (`0x0` value indicates Ether).
+      token: TokenAddress, // Address of token used for reward payout (`0x0` value indicates Ether).
     },
     {},
     ColonyClient,
@@ -554,7 +554,7 @@ export default class ColonyClient extends ContractClient {
     }>,
     TaskWorkerPayoutChanged: ContractClient.Event<{
       id: number,
-      token: Address,
+      token: TokenAddress,
       amount: number,
     }>,
     TaskFinalized: ContractClient.Event<{
@@ -641,7 +641,7 @@ export default class ColonyClient extends ContractClient {
 
     makeTaskCaller(
       'getTaskPayout',
-      [['role', 'role'], ['source', 'payableAddress']],
+      [['role', 'role'], ['token', 'tokenAddress']],
       [['amount', 'bigNumber']],
     );
     makeTaskCaller(
@@ -690,11 +690,11 @@ export default class ColonyClient extends ContractClient {
       output: [['count', 'number']],
     });
     this.addCaller('getNonRewardPotsTotal', {
-      input: [['source', 'payableAddress']],
+      input: [['token', 'tokenAddress']],
       output: [['total', 'bigNumber']],
     });
     this.addCaller('getPotBalance', {
-      input: [['potId', 'number'], ['source', 'payableAddress']],
+      input: [['potId', 'number'], ['token', 'tokenAddress']],
       output: [['balance', 'bigNumber']],
     });
     this.addCaller('getRewardPayoutInfo', {
@@ -704,7 +704,7 @@ export default class ColonyClient extends ContractClient {
         ['totalTokens', 'bigNumber'],
         ['totalTokenAmountForRewardPayout', 'bigNumber'],
         ['remainingTokenAmount', 'bigNumber'],
-        ['source', 'payableAddress'],
+        ['token', 'tokenAddress'],
         ['blockNumber', 'number'],
       ],
     });
@@ -740,7 +740,7 @@ export default class ColonyClient extends ContractClient {
     ]);
     this.addEvent('TaskWorkerPayoutChanged', [
       ['id', 'number'],
-      ['token', 'address'],
+      ['token', 'tokenAddress'],
       ['amount', 'number'],
     ]);
     this.addEvent('TaskFinalized', [['id', 'number']]);
@@ -781,13 +781,13 @@ export default class ColonyClient extends ContractClient {
       input: [['taskId', 'number']],
     });
     this.addSender('claimColonyFunds', {
-      input: [['source', 'payableAddress']],
+      input: [['token', 'tokenAddress']],
     });
     this.addSender('claimPayout', {
       input: [
         ['taskId', 'number'],
         ['role', 'role'],
-        ['source', 'payableAddress'],
+        ['token', 'tokenAddress'],
       ],
     });
     this.addSender('createTask', {
@@ -824,7 +824,7 @@ export default class ColonyClient extends ContractClient {
         ['fromPot', 'number'],
         ['toPot', 'number'],
         ['amount', 'bigNumber'],
-        ['address', 'address'],
+        ['token', 'tokenAddress'],
       ],
     });
     this.addSender('revealTaskWorkRating', {
@@ -844,7 +844,7 @@ export default class ColonyClient extends ContractClient {
     this.addSender('setTaskManagerPayout', {
       input: [
         ['taskId', 'number'],
-        ['source', 'payableAddress'],
+        ['token', 'tokenAddress'],
         ['amount', 'bigNumber'],
       ],
     });
@@ -855,7 +855,7 @@ export default class ColonyClient extends ContractClient {
       input: [['taskId', 'number'], ['deliverableHash', 'ipfsHash']],
     });
     this.addSender('startNextRewardPayout', {
-      input: [['token', 'address']],
+      input: [['token', 'tokenAddress']],
     });
     this.addSender('waiveRewardPayouts', {
       input: [['numPayouts', 'number']],
@@ -898,12 +898,12 @@ export default class ColonyClient extends ContractClient {
     );
     makeExecuteTaskChange(
       'setTaskWorkerPayout',
-      [['source', 'payableAddress'], ['amount', 'bigNumber']],
+      [['token', 'tokenAddress'], ['amount', 'bigNumber']],
       [MANAGER_ROLE, WORKER_ROLE],
     );
     makeExecuteTaskChange(
       'setTaskEvaluatorPayout',
-      [['source', 'payableAddress'], ['amount', 'bigNumber']],
+      [['token', 'tokenAddress'], ['amount', 'bigNumber']],
       [MANAGER_ROLE, EVALUATOR_ROLE],
     );
   }